refactor: 로직 폴더링
