+++ src/main/kotlin/lotto/domain/LottoFactory.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoFactory {
     fun generateLottoNumbers(): List<Int> {
