+++ src/main/kotlin/lotto/domain/Purchase.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class Purchase(
     val amountOfPurchase: String,
