+++ src/test/kotlin/lotto/LottoFactoryTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.domain.LottoFactory
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
